+++ src/main/java/org/project/bilryozo/global/security/login/dto/LoginRequestDto.java
@@ -1,4 +1,4 @@
-package org.project.bilryozo.domain.users.dto.request;
+package org.project.bilryozo.global.security.login.dto;
 
 import lombok.Getter;
 import lombok.Setter;
